django/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
cachetools==5.3.2
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
defusedxml==0.7.1
distlib==0.3.7
dj-database-url==2.1.0
dj-rest-auth==5.0.1
Django==4.2.7
django-allauth==0.55.2
django-axes==6.1.1
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.13.1
flake8==6.1.0
flower==2.0.1
gitdb==4.0.11
GitPython==3.1.40
honcho==1.1.0
humanize==4.8.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.3
mccabe==0.7.0
msgpack==1.0.7
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
prometheus-client==0.18.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pymemcache==4.0.0
pyOpenSSL==23.3.0
pyproject-api==1.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
tornado==6.3.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
virtualenv==20.24.7
wcwidth==0.2.10
whitenoise==6.6.0
zope.interface==6.1
django-crum==0.7.9